Refactor class.
Issue: IBP-5047
Review: none
